Fixed warnings and tests

+++ src/utils/RingBuffer.hpp
@@ -79,8 +79,8 @@ class RingBuffer {
      * @return false 
      */
     bool isPreviousHalfValid() const {
-        auto point1 = validParts.begin() + ((headPart - FinnUtils::ceil(parts/2.0F)) % parts);
-        auto point2 = validParts.begin() + (headPart % parts);
+        auto point1 = validParts.begin() + static_cast<long int>((headPart - static_cast<size_t>(FinnUtils::ceil(static_cast<float>(parts) / 2.0F))) % parts);
+        auto point2 = validParts.begin() + static_cast<long int>((headPart % parts));
         return std::all_of(std::min(point1, point2), std::max(point1, point2), [](bool i){return i;});
     }
